Feat: 로또 번호를 오름차순으로 정렬하여 출력
